Made generic `DataTable` component
`PhaseTable` now uses a `DataTable` internally.
In the future, other table pages can be refactored to
use this reusable component.
